Fix issue#001 | renamed aux.py -> funcs.py.
