+++ app/app.py
@@ -5,7 +5,7 @@ from tkinter import ttk
 from tabs.balance_sheet_tab import BalanceSheetTab
 from tabs.expenses_tab import ExpenseTab
 from tabs.income_statement_tab import IncomeStatementTab
-from tabs.inventory_tab import *
+from tabs.inventory_tab import InventoryTab
 from tabs.settings_tab import SettingsTab
 from tabs.transactions_tab import TransactionsTab
 
@@ -17,6 +17,8 @@ from classes.business_class import Business
 
 
 def main():
+
+
     app = SmallBusinessManager()
     app.geometry("1200x800")
     app.resizable(width=False, height=False)
