+++ app/tabs/balance_sheet_tab.py
@@ -3,7 +3,7 @@ import tkinter as tk
 from tkinter import ttk
 
 import datetime
-from classes.aux import *
+from classes.funcs import *
 
 class BalanceSheetTab(ttk.Frame):
     def __init__(self, parent: tk.Tk | ttk.Notebook, database, BUSINESS, income_statement_tab):
