+++ app/tabs/income_statement_tab.py
@@ -2,7 +2,7 @@ import tkinter as tk
 
 from tkinter import ttk
 
-from classes.aux import *
+from classes.funcs import *
 
 class IncomeStatementTab(ttk.Frame):
     def __init__(self, parent: tk.Tk | ttk.Notebook, text: str, database, BUSINESS):
